Add files via upload
This animation uses user imput in the form of the press of a left or right key to move a line, which can reflect a ball. On the opponents side is a randomly generated defense. If the ball hits a wall, defense, or paddle the ball will bounce back. The goal is to score in the goal  and you will receive a point for doing so. The overall score as well as the success state of the last point you scored are displayed in the middle of the board.